Bug fixes
Fixed the Game Over issue.
