--- Project1/nbproject/project.properties
@@ -1,1 +1,0 @@
-#Tue Mar 26 17:49:26 PDT 2024
